Struct egg::raw::RawEGraph

pub struct RawEGraph<L: Language, D, U = ()> { /* private fields */ }

A data structure to keep track of equalities between expressions.

Check out the background tutorial for more information on e-graphs in general.

E-graphs in egg::raw

In egg::raw, the main types associated with e-graphs are RawEGraph, RawEClass, Language, and Id.

RawEGraph and RawEClass are all generic over a Language, meaning that types actually floating around in the egraph are all user-defined. In particular, the e-nodes are elements of your Language. RawEGraphs and RawEClasses are additionally parameterized by some abritrary data associated with each e-class.

Many methods of RawEGraph deal with Ids, which represent e-classes. Because eclasses are frequently merged, many Ids will refer to the same e-class.

RawEGraph provides a low level API for dealing with egraphs, in particular with handling the data stored in each RawEClass so user will likely want to implemented wrappers around raw_add, raw_union, and raw_rebuild to properly handle this data *

Implementations

impl<L: Language, D, U> RawEGraph<L, D, U>

pub fn classes(&self) -> impl ExactSizeIterator<Item = &RawEClass<D>>

Returns an iterator over the eclasses in the egraph.

pub fn classes_mut(
&mut self
) -> (impl ExactSizeIterator<Item = &mut RawEClass<D>>, &mut EGraphResidual<L>)

Returns a mutating iterator over the eclasses in the egraph. Also returns the EGraphResidual so it can still be used while self is borrowed

pub fn number_of_classes(&self) -> usize

Returns the number of eclasses in the egraph.

pub fn get_class<I: BorrowMut<Id>>(&self, id: I) -> &RawEClass<D>

Returns the eclass corresponding to id

pub fn get_class_with_cannon(&self, id: Id) -> &RawEClass<D>

Like get_class but panics if id is not canonical

pub fn get_class_mut<I: BorrowMut<Id>>(
&mut self,
id: I
) -> (&mut RawEClass<D>, &mut EGraphResidual<L>)

Returns the eclass corresponding to id Also returns the EGraphResidual so it can still be used while self is borrowed

pub fn get_class_mut_with_cannon(
&mut self,
id: Id
) -> (&mut RawEClass<D>, &mut EGraphResidual<L>)

Like get_class_mut but panics if id is not canonical

pub fn is_clean(&self) -> bool

Returns whether self is congruently closed

This will always be true after calling raw_rebuild

impl<L: Language, D, U: UndoLogT<L, D>> RawEGraph<L, D, U>

pub fn raw_add<T>(
outer: &mut T,
get_self: impl Fn(&mut T) -> &mut Self,
original: L,
handle_equiv: impl FnOnce(&mut T, Id, &L) -> Option<Id>,
handle_union: impl FnOnce(&mut T, Id, Id),
mk_data: impl FnOnce(&mut T, Id, &L) -> D
) -> Id

Adds enode to a RawEGraph contained within a wrapper type T

Parameters

get_self

Called to extract the RawEGraph from the wrapper type, and should not perform any mutation.

This will likely be a simple field access or just the identity function if there is no wrapper type.

handle_equiv

When there already exists a node that is congruently equivalent to enode in the egraph this function is called with the uncanonical id of a equivalent node, and a reference to enode

Returning Some(id) will cause raw_add to immediately return id (in this case id should represent an enode that is equivalent to the one being inserted).

Returning None will cause raw_add to create a new id for enode, union it to the equivalent node, and then return it.

handle_union

Called after handle_equiv returns None with the uncanonical id of the equivalent node and the new id assigned to enode

Calling id_to_node on the new id will return a reference to enode

mk_data

When there does not already exist a node is congruently equivalent to enode in the egraph this function is called with the new id assigned to enode and a reference to the canonicalized version of enode to create to data that will be stored in the RawEClass associated with it

Calling id_to_node on the new id will return a reference to enode

Calling get_class on the new id will cause a panic since the RawEClass is still being built

pub fn raw_add_for_sym<T, X>(
outer: &mut T,
get_self: impl Fn(&mut T) -> &mut Self,
original: L,
handle_equiv: impl FnOnce(&mut T, Id, &L) -> Option<Id>,
pre_union: impl FnOnce(&L, &L) -> X,
handle_union: impl FnOnce(&mut T, X, Id, Id),
mk_data: impl FnOnce(&mut T, Id, &L) -> D
) -> Id

Similar to raw_add but takes an extra pre_union closure that allows comparing the two equal canonical nodes used to determine a congruent union. This can be useful in niche situations involving explanations when node canonization sorts children to make some discriminants symmetric

Note pre_union is called before handle_equiv even though it is only relevant when handle_equiv returns None

pub fn raw_union(
&mut self,
enode_id1: Id,
enode_id2: Id,
merge: impl FnOnce(MergeInfo<'_, D>)
)

Unions two eclasses given their ids.

The given ids need not be canonical.

If a union occurs, merge is called with the data, id, and parents of the two eclasses being merged

pub fn raw_rebuild<T>(
outer: &mut T,
get_self: impl Fn(&mut T) -> &mut Self,
perform_union: impl FnMut(&mut T, Id, Id),
handle_pending: impl FnMut(&mut T, Id, &L)
)

Rebuild to RawEGraph to restore congruence closure

Parameters

get_self

Called to extract the RawEGraph from the wrapper type, and should not perform any mutation.

This will likely be a simple field access or just the identity function if there is no wrapper type.

perform_union

Called on each pair of ids that needs to be unioned

In order to be correct perform_union should call raw_union

handle_pending

Called with the uncanonical id of each enode whose canonical children have changed, along with a canonical version of it

pub fn try_raw_rebuild<T, E>(
outer: &mut T,
get_self: impl Fn(&mut T) -> &mut Self,
perform_union: impl FnMut(&mut T, Id, Id) -> Result<(), E>,
handle_pending: impl FnMut(&mut T, Id, &L)
) -> Result<(), E>

Similar to [raw_rebuild] but allows for the union operation to fail and abort the rebuild

pub fn try_raw_rebuild_for_sym<T, E, X>(
outer: &mut T,
get_self: impl Fn(&mut T) -> &mut Self,
pre_union: impl FnMut(&L, &L) -> X,
perform_union: impl FnMut(&mut T, X, Id, Id) -> Result<(), E>,
handle_pending: impl FnMut(&mut T, Id, &L)
) -> Result<(), E>

Similar to try_raw_rebuild but takes an extra pre_union closure that allows comparing the two equal canonical nodes used to determine a congruent union. This can be useful in niche situations involving explanations when node canonization sorts children to make some discriminants symmetric

pub fn dump_classes(&self) -> impl Debug + '_ where
D: Debug,

Returns a more debug-able representation of the egraph focusing on its classes.

RawEGraphs implement Debug, but it’s not pretty. It prints a lot of stuff you probably don’t care about. This method returns a wrapper that implements Debug in a slightly nicer way, just dumping enodes in each eclass.

pub fn clear(&mut self)

Remove all nodes from this egraph

impl<L: Language, U: UndoLogT<L, ()>> RawEGraph<L, (), U>

pub fn add_uncanonical(&mut self, enode: L) -> Id

Simplified version of raw_add for egraphs without eclass data

pub fn union(&mut self, id1: Id, id2: Id) -> bool

Simplified version of raw_union for egraphs without eclass data

pub fn rebuild(&mut self)

Simplified version of raw_rebuild for egraphs without eclass data

impl<L: Language, D, U: UndoLogT<L, D>> RawEGraph<L, D, U>

pub fn set_undo_log(&mut self, undo_log: U)

Change the UndoLogT being used

If the new UndoLogT is enabled then the egraph must be empty

pub fn has_undo_log(&self) -> bool

Check if the UndoLogT being used is enabled

impl<L: Language, D, U: AsUnwrap<UndoLog>> RawEGraph<L, D, U>

pub fn push1(&self) -> PushInfo

Create a PushInfo representing the current state of the egraph which can later be passed into raw_pop1

Requires self.is_clean()

Example

use egg::{raw::*, SymbolLang as S};
use egg::raw::semi_persistent1::UndoLog;
let mut egraph = RawEGraph::<S, (), UndoLog>::default();
let a = egraph.add_uncanonical(S::leaf("a"));
let fa = egraph.add_uncanonical(S::new("f", vec![a]));
let c = egraph.add_uncanonical(S::leaf("c"));
egraph.rebuild();
let old = egraph.clone();
let restore_point = egraph.push1();
let b = egraph.add_uncanonical(S::leaf("b"));
let _fb = egraph.add_uncanonical(S::new("g", vec![b]));
egraph.union(b, a);
egraph.union(b, c);
egraph.rebuild();
egraph.pop1(restore_point);
assert_eq!(format!("{:#?}", egraph.dump_uncanonical()), format!("{:#?}", old.dump_uncanonical()));
assert_eq!(format!("{:#?}", egraph), format!("{:#?}", old));

pub fn raw_pop1(
&mut self,
info: PushInfo,
split: impl FnMut(&mut D, Id, Id) -> D
)

Mostly restores the egraph to the state it was it when it called push1 to create info

Invalidates all PushInfos that were created after info

The raw_data fields of the RawEClasss are not properly restored Instead, split is called to undo each union with a mutable reference to the merged data, and the two ids that were merged to create the data for the eclass of the second id (the eclass of the first id will be what’s left of the merged data after the call)

pub fn find_direct_parent(&self, id: Id) -> Id

Return the direct parent from the union find without path compression

impl<L: Language, U: AsUnwrap<UndoLog>> RawEGraph<L, (), U>

pub fn pop1(&mut self, info: PushInfo)

Simplified version of raw_pop1 for egraphs without eclass data

impl<L: Language, D, U: AsUnwrap<UndoLog>> RawEGraph<L, D, U>

pub fn push2(&self) -> PushInfo

Create a PushInfo representing the current state of the egraph which can later be passed into raw_pop2

Requires self.is_clean()

Example

use egg::{raw::*, SymbolLang as S};
use egg::raw::semi_persistent2::UndoLog;
let mut egraph = RawEGraph::<S, (), UndoLog>::default();
let a = egraph.add_uncanonical(S::leaf("a"));
let fa = egraph.add_uncanonical(S::new("f", vec![a]));
let c = egraph.add_uncanonical(S::leaf("c"));
egraph.rebuild();
assert_eq!(egraph.number_of_classes(), 3);
assert_eq!(egraph.number_of_uncanonical_nodes(), 3);
assert_eq!(egraph.total_size(), 3);
let restore_point = egraph.push2();
let b = egraph.add_uncanonical(S::leaf("b"));
let _fb = egraph.add_uncanonical(S::new("g", vec![b]));
egraph.union(b, a);
egraph.union(b, c);
egraph.rebuild();
assert_eq!(egraph.find(a), b);
assert_eq!(egraph.number_of_classes(), 3);
assert_eq!(egraph.number_of_uncanonical_nodes(), 5);
assert_eq!(egraph.total_size(), 6);
egraph.pop2(restore_point);
assert_ne!(egraph.find(a), b);
assert_eq!(egraph.lookup(S::leaf("a")), Some(a));
assert_eq!(egraph.lookup(S::new("f", vec![a])), Some(fa));
assert_eq!(egraph.lookup(S::leaf("b")), None);
assert_eq!(egraph.number_of_classes(), 3);
assert_eq!(egraph.number_of_uncanonical_nodes(), 3);
assert_eq!(egraph.total_size(), 3);

pub fn raw_pop2<T>(
&mut self,
info: PushInfo,
state: &mut T,
clear: impl FnMut(&mut T, &mut D, Id, UndoCtx<'_, L>),
mk_data: impl FnMut(&mut T, Id, UndoCtx<'_, L>) -> D,
handle_eqv: impl FnMut(&mut T, &mut D, Id, UndoCtx<'_, L>)
)

Mostly restores the egraph to the state it was it when it called push2 to create info

Invalidates all PushInfos that were created after info

The raw_data fields of the RawEClasss are not properly restored Instead all eclasses that have were merged into another eclass are recreated with mk_data and clear is called eclass that had another eclass merged into them

After each call to either mk_data or clear, handle_eqv is called on each id that is in the eclass (that was handled by mk_data or clear

The state parameter represents arbitrary state that be accessed in any of the closures

pub fn undo_ctx(&self) -> UndoCtx<'_, L>

Returns the UndoCtx corresponding to the current egraph

impl<L: Language, U: AsUnwrap<UndoLog>> RawEGraph<L, (), U>

pub fn pop2(&mut self, info: PushInfo)

Simplified version of raw_pop2 for egraphs without eclass data