Struct egg::ConditionEqual

pub struct ConditionEqual<L> { /* private fields */ }

A Condition that checks if two terms are equivalent.

This condition adds its two Pattern to the egraph and passes if and only if they are equivalent (in the same eclass).

Implementations

impl<L: Language> ConditionEqual<L>

pub fn new(p1: Pattern<L>, p2: Pattern<L>) -> Self

Create a new ConditionEqual condition given two patterns.

impl<L: FromOp> ConditionEqual<L>

pub fn parse(a1: &str, a2: &str) -> Self

Create a ConditionEqual by parsing two pattern strings.

This panics if the parsing fails.