egg

Struct EClass

#[non_exhaustive]
pub struct EClass<L, D> {
    pub id: Id,
    pub nodes: Vec<L>,
    pub data: D,
    /* private fields */
}

An equivalence class of enodes.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

id: Id

This eclass’s id.

nodes: Vec<L>

The equivalent enodes in this equivalence class.

data: D

The analysis data associated with this eclass.

Modifying this field will not cause changes to propagate through the e-graph. Prefer EGraph::set_analysis_data instead.

Implementations

impl<L, D> EClass<L, D>

pub fn is_empty(&self) -> bool

Returns true if the eclass is empty.

pub fn len(&self) -> usize

Returns the number of enodes in this eclass.

pub fn iter(&self) -> impl ExactSizeIterator<Item = &L>

Iterates over the enodes in this eclass.

pub fn parents(&self) -> impl ExactSizeIterator<Item = Id> + '_

Iterates over the non-canonical ids of parent enodes of this eclass.

impl<L: Language, D> EClass<L, D>

pub fn leaves(&self) -> impl Iterator<Item = &L>

Iterates over the childless enodes in this eclass.

pub fn assert_unique_leaves(&self)
where L: Language,

Asserts that the childless enodes in this eclass are unique.

pub fn for_each_matching_node<Err>( &self, node: &L, f: impl FnMut(&L) -> Result<(), Err>, ) -> Result<(), Err>
where L: Language,

Run some function on each matching e-node in this class.