Struct egg::Symbol

pub struct Symbol(/* private fields */);

An interned string.

This is provided by the symbol_table crate.

Internally, egg frequently compares Vars and elements of Languages. To keep comparisons fast, egg provides Symbol a simple wrapper providing interned strings.

You may wish to use Symbol in your own Languages to increase performance and keep enode sizes down (a Symbol is only 4 bytes, compared to 24 for a String.)

A Symbol is simply a wrapper around an integer. When creating a Symbol from a string, egg looks up it up in a global table, returning the index (inserting it if not found). That integer is used to cheaply implement Copy, Clone, PartialEq, Eq, PartialOrd, Ord, and Hash.

The internal symbol cache leaks the strings, which should be fine if you only put in things like variable names and identifiers.

Example

use egg::Symbol;

assert_eq!(Symbol::from("foo"), Symbol::from("foo"));
assert_eq!(Symbol::from("foo"), "foo".parse().unwrap());

assert_ne!(Symbol::from("foo"), Symbol::from("bar"));

A interned string in the global symbol table.

This requires the global feature on the crate.

GlobalSymbol is a wrapper around Symbol that knows to refer to a built-in, global [SymbolTable]. Strings into the global table are never freed.

This enables a lot of convenience methods and trait implementations over GlobalSymbol (see below). In particular, you can convert it to &'static str, convert From and Into a &str, and de/serialize using serde if the serde feature is enabled.

Implementations

impl GlobalSymbol

pub fn new(s: impl AsRef<str>) -> GlobalSymbol

Intern a string into the global symbol table.

pub fn as_str(&self) -> &'static str

Convert this symbol into the string in the static, global symbol table.