Trait egg::IterationData

pub trait IterationData<L, N>: Sized
where
    L: Language,
    N: Analysis<L>,
{
    // Required method
    fn make(runner: &Runner<L, N, Self>) -> Self;
}

Custom data to inject into the Iterations recorded by a Runner

This trait allows you to add custom data to the Iterations recorded as a Runner applies rules.

See the Runner docs for an example.

Runner is generic over the IterationData that it will be in the Iterations, but by default it uses ().

Required Methods

fn make(runner: &Runner<L, N, Self>) -> Self

Given the current Runner, make the data to be put in this Iteration.

Object Safety

This trait is not object safe.