egg

Trait LpCostFunction

pub trait LpCostFunction<L: Language, N: Analysis<L>> {
    // Required method
    fn node_cost(&mut self, egraph: &EGraph<L, N>, eclass: Id, enode: &L) -> f64;
}

A cost function to be used by an LpExtractor.

Required Methods

fn node_cost(&mut self, egraph: &EGraph<L, N>, eclass: Id, enode: &L) -> f64

Returns the cost of the given e-node.

This function may look at other parts of the e-graph to compute the cost of the given e-node.