Crate egg

egg (e-graphs good) is a e-graph library optimized for equality saturation.

This is the API documentation.

The tutorial is a good starting point if you're new to e-graphs, equality saturation, or Rust.

The tests on Github provide some more elaborate examples.

There is also a paper describing egg and some of its technical novelties.

Logging

Many parts of egg dump useful logging info using the log crate. The easiest way to see this info is to use the env_logger crate in your binary or test. The simplest way to enable env_logger is to put the following line near the top of your main: env_logger::init();. Then, set the environment variable RUST_LOG=egg=info, or use warn or debug instead of info for less or more logging.

Modules

tutorials
A Guide-level Explanation of egg

Macros

define_language
A macro to easily create a Language.

rewrite
A macro to easily make Rewrites.

test_fn
Make a test function

Structs

AstDepth
A simple CostFunction that counts maximum ast depth.

AstSize
A simple CostFunction that counts total ast size.

BackoffScheduler
A RewriteScheduler that implements exponentional rule backoff.

ConditionEqual
A Condition that checks if two terms are equivalent.

ConditionalApplier
An Applier that checks a Condition before applying.

DidMerge
Result of Analysis::merge indicating which of the inputs are different from the merged result.

Dot
A wrapper for an EGraph that can output GraphViz for visualization.

EClass
An equivalence class of enodes.

EGraph
A data structure to keep track of equalities between expressions.

Explanation
A data structure representing an explanation that two terms are equivalent.

Extractor
Extracting a single RecExpr from an EGraph.

FlatTerm
A single term in an flattened explanation. After the first term in a FlatExplanation, each term will be annotated with exactly one backward_rule or one forward_rule. This can appear in children FlatTerms, indicating that the child is being rewritten.

FromOpError
A generic error for failing to parse an operator. This is the error type used by define_language! for FromOp::Error, and is a sensible choice when implementing FromOp manually.

Id
A key to identify EClasses within an EGraph.

Iteration
Data generated by running a Runner one iteration.

Pattern
A pattern that can function as either a Searcher or Applier.

RecExpr
A recursive expression from a user-defined Language.

Report
A report containing data about an entire Runner run.

Rewrite
A rewrite that searches for the lefthand side and applies the righthand side.

Runner
Faciliates running rewrites over an EGraph.

SearchMatches
The result of searching a Searcher over one eclass.

SimpleScheduler
A very simple RewriteScheduler that runs every rewrite every time.

Subst
A substitition mapping Vars to eclass Ids.

Symbol
An interned string.

SymbolLang
A simple language used for testing.

TreeTerm
An explanation for a term and its equivalent children. Each child is a proof transforming the initial child into the final child term. The initial term is given by taking each first sub-term in each child_proofs recursivly. The final term is given by all of the final terms in each child_proofs.

Var
A variable for use in Patterns or Substs.

Enums

ENodeOrVar
The language of Patterns.

RecExprParseError
An error type for failures when attempting to parse an s-expression as a RecExpr<L>.

StopReason
Error returned by Runner when it stops.

Traits

Analysis
Arbitrary data associated with an EClass.

Applier
The righthand side of a Rewrite.

Condition
A condition to check in a ConditionalApplier.

CostFunction
A cost function that can be used by an Extractor.

FromOp
A trait for parsing e-nodes. This is implemented automatically by define_language!.

IterationData
Custom data to inject into the Iterations recorded by a Runner

Language
Trait that defines a Language whose terms will be in the EGraph.

LanguageChildren
A marker that defines acceptable children types for define_language!.

RewriteScheduler
A way to customize how a Runner runs Rewrites.

Searcher
The lefthand side of a Rewrite.

Functions

merge_max
A utility for implementing Analysis::merge when the Data type has a total ordering. This will take the maximum of the two values.

merge_min
A utility for implementing Analysis::merge when the Data type has a total ordering. This will take the minimum of the two values.

Type Definitions

FlatExplanation
FlatExplanation are the simpler, expanded representation showing one term being rewritten to another. Each step contains a full FlatTerm. Each flat term is connected to the previous by exactly one rewrite.

PatternAst
A RecExpr that represents a Pattern.

TreeExplanation
Explanation trees are the compact representation showing how one term can be rewritten to another.