Trait egg::Analysis

pub trait Analysis<L: Language>: Sized {
    type Data: Debug;
    fn make(egraph: &EGraph<L, Self>, enode: &L) -> Self::Data;

    fn merge(&self, a: &mut Self::Data, b: Self::Data) -> DidMerge;

    fn pre_union(egraph: &EGraph<L, Self>, id1: Id, id2: Id) { ... }

    fn modify(egraph: &mut EGraph<L, Self>, id: Id) { ... }
}

Associated Types

type Data: Debug

The per-EClass data for this analysis.

Loading content...
Required methods

fn make(egraph: &EGraph<L, Self>, enode: &L) -> Self::Data

Makes a new Analysis for a given enode Analysis.

fn merge(&self, a: &mut Self::Data, b: Self::Data) -> DidMerge

Defines how to merge two Datas when their containing EClasses merge.

This should update a to correspond to the merged analysis data.

The result is a DidMerge(a_merged, b_merged) indicating whether the merged result is different from a and b respectively.

Since merge can modify a, let a0/a1 be the value of a before/after the call to merge, respectively.

If a0 != a1 the result must have a_merged == true. This may be conservative -- it may be true even if even if a0 == a1.

If b != a1 the result must have b_merged == true. This may be conservative -- it may be true even if even if b == a1.

Loading content...
Provided methods

fn pre_union(egraph: &EGraph<L, Self>, id1: Id, id2: Id)

An optional hook that allows inspection before a union occurs.

By default it does nothing.

pre_union is called a lot, so doing anything significant (like printing) will cause things to slow down.

fn modify(egraph: &mut EGraph<L, Self>, id: Id)

A hook that allows the modification of the EGraph

By default this does nothing.

Loading content...