Trait egg::LanguageChildren

pub trait LanguageChildren {
    // Required methods
    fn len(&self) -> usize;
    fn can_be_length(n: usize) -> bool;
    fn from_vec(v: Vec<Id>) -> Self;
    fn as_slice(&self) -> &[Id];
    fn as_mut_slice(&mut self) -> &mut [Id];

    // Provided method
    fn is_empty(&self) -> bool { ... }
}

A marker that defines acceptable children types for define_language!.

See define_language! for more details. You should not have to implement this trait.

Required Methods

fn len(&self) -> usize

Returns the number of children.

fn can_be_length(n: usize) -> bool

Checks if n is an acceptable number of children for this type.

fn from_vec(v: Vec<Id>) -> Self

Create an instance of this type from a Vec<Id>, with the guarantee that can_be_length is already true on the Vec.

fn as_slice(&self) -> &[Id]

Returns a slice of the children Ids.

fn as_mut_slice(&mut self) -> &mut [Id]

Returns a mutable slice of the children Ids.

Provided Methods

fn is_empty(&self) -> bool

Checks if there are no children.

Object Safety

This trait is not object safe.